+++ ccr_rw_map.h
@@ -6,8 +6,8 @@
 #include <stdio.h>
 #define HASH(n) ((n)%100)
 typedef struct _map_slot{
-	uint32_t key;
-	uint32_t val;
+	uint64_t key;
+	uint64_t val;
 	struct _map_slot * next;
 } slot;
 // a map that maps uint32_t (id) to uint32_t () 
@@ -19,8 +19,8 @@ typedef struct _concurrnt_rw_map{
 int ccr_rw_map_init(ccr_rw_map * cmap);
 // insert a <key,val> pair into map
 // return 0 if success -1 fail
-int ccr_rw_map_insert(ccr_rw_map * cmap,uint32_t k,uint32_t v);
+int ccr_rw_map_insert(ccr_rw_map * cmap,uint64_t k,uint64_t v);
 // query in map
 // return 0 if success -1 k not found in map -2 other errors
-int ccr_rw_map_query(ccr_rw_map * cmap,uint32_t k,uint32_t * v);
+int ccr_rw_map_query(ccr_rw_map * cmap,uint64_t k,uint64_t * v);
 #endif
